+++ doc/5_02_message.qbk
@@ -146,6 +146,8 @@ meet the requirements, or use the ones that come with the library:
 ]]
 [[
     [link beast.ref.beast__http__file_body `file_body`]
+
+    [link beast.ref.beast__http__basic_file_body `basic_file_body`]
 ][
     This body is represented by a file opened for either reading or
     writing. Messages with this body may be serialized and parsed.
