+++ doc/quickref.xml
@@ -30,14 +30,15 @@
           <bridgehead renderas="sect3">Classes</bridgehead>
           <simplelist type="vert" columns="1">
             <member><link linkend="beast.ref.beast__http__basic_dynamic_body">basic_dynamic_body</link></member>
-            <member><link linkend="beast.ref.beast__http__basic_file_body">basic_file_body</link></member>
             <member><link linkend="beast.ref.beast__http__basic_fields">basic_fields</link></member>
+            <member><link linkend="beast.ref.beast__http__basic_file_body">basic_file_body</link></member>
             <member><link linkend="beast.ref.beast__http__basic_parser">basic_parser</link></member>
             <member><link linkend="beast.ref.beast__http__basic_string_body">basic_string_body</link></member>
             <member><link linkend="beast.ref.beast__http__buffer_body">buffer_body</link></member>
             <member><link linkend="beast.ref.beast__http__dynamic_body">dynamic_body</link></member>
             <member><link linkend="beast.ref.beast__http__empty_body">empty_body</link></member>
             <member><link linkend="beast.ref.beast__http__fields">fields</link></member>
+            <member><link linkend="beast.ref.beast__http__file_body">file_body</link></member>
             <member><link linkend="beast.ref.beast__http__header">header</link></member>
             <member><link linkend="beast.ref.beast__http__message">message</link></member>
             <member><link linkend="beast.ref.beast__http__parser">parser</link></member>
