+++ test/http/Jamfile
@@ -8,6 +8,7 @@
 unit-test http-tests :
     ../../extras/beast/unit_test/main.cpp
     basic_dynamic_body.cpp
+    basic_file_body.cpp
     basic_parser.cpp
     buffer_body.cpp
     doc_examples.cpp
